HOT FISH

+++ src/models/ebau_tests.ts
@@ -10,9 +10,9 @@ export const EbauTest = sequelize.define('EBAU_PRUEBA', {
   CURSO_ACA: DataTypes.STRING,
   CONV: DataTypes.STRING,
   ESPECIALIDAD: DataTypes.STRING,
-  NOTA_BACH: DataTypes.INTEGER,
-  NOTA_PRUE:  DataTypes.INTEGER,
-  NOTA_DEF: DataTypes.INTEGER
+  NOTA_BACH: DataTypes.FLOAT,
+  NOTA_PRUE:  DataTypes.FLOAT,
+  NOTA_DEF: DataTypes.FLOAT
 }, {
   freezeTableName: true,
   timestamps: false
